Introduce zeroOrOneResult argument to methods that mask flags fields
Various methods generate IL to mask the flags or classDepthAndFlags
fields.  In some cases, IL that tests the result of the masking is
simply interested in whether it is zero or non-zero.  However, in a few
cases the wants the result to be zero or one.

If the mask was a power of two, the result of the masking is shifted
right to place the masked bit in the least significant bit position.
Otherwise, the result of the masking is compared with zero using
TR::icmpne to yield a result of zero or one.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>