Use custom allocator for JITServer client MessageBuffers
A global, reference-counted allocator is now created at JITServer
clients that is used to allocate the memory used by the internal
MessageBuffer storage buffer during client/server communication.
This allocator's memory segments are disclaimed if the number of
connections to the server drops to zero, and the compiler will also
periodically attempt to destroy the allocator entirely.

Signed-off-by: Christian Despres <despresc@ibm.com>