Clarify naming of variables for paths of a diamond
Various places in the inliner refer to the target TR::Block,
TR::TreeTop or TR::Node of a branch in the IL as the "if" path and the
fall-through path as the "else" path.  Programmers might have different
intuitions about whether the fall-through path is the "if path" or the
"else" path.

To avoid confusion, this change updates variable names to make it clear
whether they represent things that will be executed if the branch is
taken or executed if the branch is not taken.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>