Reduce number of VM_getClassFromSignature messages for JITServer
VM_getClassFromSignature messages account for 10-11% of all
messages exchanged between client and server. Such messages
are sent when the server calls getClassFromSignature() to
obtain a j9class pointer starting from a class name.
This commit is based on the observation that primitive types,
like I, Z, B, etc., do not have an associate j9class, so the
answer from the getClassFromSignature() query is always NULL.
This change reduces the number of VM_getClassFromSignature
messages by 33-35%

Signed-off-by: Marius <mpirvu@ca.ibm.com>