Ignore ByteBufferHandle operations in unsafe fast path
Transforming unsafe accesses to array accesses when a ByteBufferHandle
operation method is on the inline call stack erroneously allows a read-
only buffer to be mutated on big-endian systems. The indexRO() method
checks for read-only buffers by loading the ByteBuffer.isReadOnly field
with Unsafe. That's a boolean, which in an array would occupy one byte,
but in a field it occupies four. Treating the load as an array access
means that we only load the first byte, which on a big-endian system is
the high byte and therefore always zero.