Update Thread.interrupted in JVMTI RawMonitorWait
JVMTI GetThreadState uses Thread.interrupted to interpret the
thread's state. If a thread is interrupted while waiting for
a monitor in JVMTI RawMonitorWait, then it's Thread.interrupted
field needs to be updated. This allows JVMTI GetThreadState to
return the correct thread state.

Use JVMTI_ERROR_INTERNAL (universal error) for the CRIU error
case.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>