Fix int overflow check
The check here is supposed to guard against the UB of casting a
floating-point value outside the int32_t range to an int32_t. The value
0x7fffffff (decimal 2147483647) is not exactly representable as a 32-bit
float; both gcc and clang convert it to the float value
2147483648.0 (one more than the literal). The added explicit `(float)`
cast eliminates a clang warning about the issue of inexact
representation.

The check is also changed to `>=`, as in light of the above the previous
code would fail if `weight` were ever exactly the floating-point value
2147483648.0.

Signed-off-by: Christian Despres <despresc@ibm.com>