Fixes to code cache disclaiming
Existing code disclaims the cold code between
`coldCodeAlloc` and `coldCodeAllocBase`.
However, if the amount of cold code is larger than the
amount of warm code, we will ask to disclaim some parts
of the segment which was allocated with THP. This will
disable the THP setting and lower throughput.
This commit also introduces two new code cache fields
that are used to memorize the start and end of the section
forced to use small pages and that should be subject
to disclaiming.

Signe-off-by: Marius <mpirvu@ca.ibm.com>