Clarify which branches should be compared
The description of coordinated merges in the Guide for Committers
indicates that the tips of the master and openj9 branches in the
openj9-omr repository should be compared to avoid promoting changes that
have not passed an OMR Acceptance build.  However, even if the tips of
those branches are the same, it might be that the tip of the master
branch in the omr repository does not match the tip of the master branch
in the openj9-omr repository.  Performing a coordinated merge in that
circumstance would also promote changes that have not passed an OMR
acceptance build.

Added that case to the list of branches to compare.

[skip ci]

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>