Reduce unnecessary TLS usage in JITServer code
Avoid using the thread-local variable TR::compInfoPT (mostly through
TR::CompilationInfo::getStream()) in contexts where the compInfoPT or
the Compilation is readily available. TLS accesses in dynamically loaded
shared libraries (such as the JIT) can have non-negligible overhead.

A resubmission of #14931. That PR also avoided using TLS in the methods
of the J9::KnownObjectTable by accessing that class's comp() object to
retrieve a JITServer::ServerStream. However, that is (not at all
obviously) unsafe to do in that context, because some methods of the
J9::KnownObjectTable may be called during the initialization of that
very comp() object.

Fixes: #19809
Signed-off-by: Christian Despres <despresc@ibm.com>