Log local-following-remote compilations
If the JIT option enableJITServerFollowRemoteCompileWithLocalCompile is
enabled for a particular method, a JITServer client will compile that
method again, locally, whenever it receives a compiled version of the
method from the server. The start and end of this compilation are now
noted in the verbose log if TR_VerboseJITServer is enabled.

Signed-off-by: Christian Despres <despresc@ibm.com>