Remove requirement for consistent memory page size in SCC
1. The page level operation in SCC are msync and mprotect. msync is only
used on z/OS for persistent cache, we can always do msync if we have a
valid _osPageSize. For mprotect, it is turned off if the SCC is
readonly. If the data is not paged aligned anymore, mprotect should be
also be turned off. There is no need to force consistent memory page
size between cold and warn runs anymore.

2. Add new test flags and command line test cases.

3. Add top layer page size infomation in cache stats.

4. Add page size info into javacore.

Fixes #19968

Signed-off-by: Hang Shao <hangshao@ca.ibm.com>