Remove requirement for consistent page size in SCC
The page level operation in SCC are msync and mprotect.
msync is only used on z/OS, we can always do msync if we can have valid
_osPageSize.
For mprotect, it is turned off if the SCC is readonly. If the data is
not paged aligned anymore, mprotect should be also be turned off. There
is no need to force consistent memory page size anymore.

Fixes #19968

Signed-off-by: Hang Shao <hangshao@ca.ibm.com>