Fall back to out-of-line call to StringUTF16.toBytes
The implementation of StringUTF16.toBytes will throw a
NegativeArraySizeException if the specified array length is negative or
an OutOfMemoryError if the length is 2^30 or greater.  The IL for the
inline implementation that the JIT compiler generates for that
recognized method could throw a NegativeArraySizeException in situations
where the actual implementation would throw an OutOfMemoryError.

This change calls the actual implementation if the array length is
outside the range [0,0x3fffffff] to ensure that the expected behaviour
always results in exceptional situations.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>