Make IProfiler race conditon less likely
Interpreter profiling is uses lockless mechanisms that
allow for some races to occur. The outcome of these races
is some small memory leak.
When adding new profiling info to the IProfiler hashtable, we
first determine if an entry with the same pc exists (pc is the key)
and if not, we allocate memory for a new entry and then add the
entry to the head of the linked list of the corresponding bucket
of the IP hashtable. It turns out that allocating memory for a
new entry is relatively expensive, and in this window of time
another thread could perform the same check, deciding to add
a new entry with the same pc.
This commit performs another check just before adding to the
linked list of the bucket. This reduces the likelihood of
two threads adding entries with the same pc.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>