Delay JVM phase change post restore with -XX:+DebugOnRestore
The JIT's startup heuristics have been highly tuned for a normal JVM
run. However, a checkpoint/restore run does not play nice with these
heuristics. As a consequence, the phase change occurs too early on
restore, which can cause an increase on RSS footprint because the JIT
will touch J9ROMClasses in the SCC that were disclaimed automatically
during checkpoint.

This commit adds a delay so that the JVM Phase change heuristics are
more in line with the normal JVM run.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>