Use findOrFabricateShadowSymbol for Reference.referent in get/refersTo
When inlining Reference.getImpl() or Reference.refersTo(), the compiler
generates a load of Reference.referent. Previously, it would reuse any
address-typed shadow symref it could find with the expected offset and
owning method. If it did find such a shadow symref representing a field
of a different class, it would have incorrect aliasing and likely an
incorrect type signature. Now instead the field shadow is fabricated if
necessary using findOrFabricateShadowSymbol().