Disable TLH prefetching for portable AOT code
Current code for x86 disables TLH prefething globally for
Intel CPUs newer than Broadwell architecture and enables it
globally for Broadwell or older CPUs.
If a container image is generated on a newer architecture,
then TLH prefetch will be off and this information is written
into the AOT header for the shared class cache embedded in the
container. If that container image is run on an older architecture,
the JVM will set TLH prefetch off, and the AOT compatibility
check will fail, rendering the embedded AOT code useless.

This commit disables TLH prefething for portable AOT code.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>