Remove outdated comment and logic from refineMethodHandleLinkTo()
It's true that final methods of Object and private methods are not in
the vtable and must be dispatched directly even though they're called
using invokevirtual. However, those methods won't be encountered here
because the MemberName would be of kind MH_REF_INVOKESPECIAL, and the
dispatch would be carried out by linkToSpecial().