Delete unhelpful vTableSlot assert
The TR_ResolvedJ9Method::vTableSlot() would formerly assert that the
_vTableSlot of the resolved method was not zero before returning its
value. This is not particularly helpful, as none of the existing code
assumes that the vTableSlot() will be valid when calling this method,
and there are certain kinds of resolved methods that may legitimately
have a _vTableSlot of zero.

This assert would (unnecessarily) fire in particular in the handling of
the ResolvedMethod_getMultipleResolvedMethods message at a JITServer
client. Removing it eliminates the resulting crash in
debug/prod-with-assumes builds.

Signed-off-by: Christian Despres <despresc@ibm.com>