Fix JFR timestamps
Set the ticks/second parameter in the chunk header. This indicates which
granularity we report timestamps (millis, micros or nanos). Note, this
is not to be confused with the event startime which is always
milliseconds.

Also, chage threadsleep event to report on after the sleep as concluded
which matches RI behaviour. The actual sleep duration must be recorded.

Signed-off-by: tajila <atobia@ca.ibm.com>