Return early from isFSDNeeded under -XX:+DebugOnRestore
Under -XX:+DebugOnRestore, the VM will enable some capabilities that
will be reset on restore (if the user does not want debug on restore).
However, if the JIT detects these capabilities via the J9HookDisable
call, there is no way for the JIT to know whether the capability was set
by the VM or a user pre-checkpoint.

With the recent change to the isDebugOnRestoreEnabled to check if a uesr
specified an JDWP agent pre-checkpoint, the JIT can simply return early
from isFSDNeeded. Under -XX:+DebugOnRestore, the JIT already generates
FSD code so there is on need to check the caps that were set by the VM
pre-checkpoint.

isFSDNeeded is called on restore, so at that point it is appropriate for
the JIT to execute the rest of the method at that point.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>