Adjust bound for out-of-line call to StringUTF16.toBytes
Recognized Call Transformer generates IL that implements an inline
version of StringUTF16.toBytes for values in the range [0,0x3fffffff],
and that falls back to calling the Java implementation otherwise.  In
JDK levels up to 22, the Java implementation throws an
OutOfMemoryException if the length is greater than 0x3fffffff.  JDK
Level 23, however, was changed to throw an OutOfMemoryException if the
length is greater than or equal to 0x3fffffff.

This change alters the IL that Recognized Call Transformer generates to
fall back to the Java implementation if the length is greater than
or equal to 0x3fffffff, so that it will work with all JDK levels.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>