Allow SVM assertions to be ignored
If the option TR_IgnoreAssert is enabled for a particular compilation
and the environment variable TR_svmAssertionsAreFatal is not set, SVM
assertions will now be non-fatal even in DEBUG or PROD_WITH_ASSUMES
builds. This matches the behaviour of normal TR_ASSERT assertions, aside
from the possibility of overriding TR_IgnoreAssert with that environment
variable.

Signed-off-by: Christian Despres <despresc@ibm.com>