Wait up to 1 minute in JITServer test
Part of the teardown process in the JITServer tests involves waiting for
the server and client to shut down in an orderly fashion. Occasionally
this can (legitimately) take longer than 30s for the client, the
previous maximum. This causes the test to forcibly destroy the process
and mark the test as having failed. We now wait for up to one minute
for these processes to finish.

Signed-off-by: Christian Despres <despresc@ibm.com>