Stop setting known object index in findOrCreateStaticSymbol()
SymbolReferenceTable::findOrCreateStaticSymbol() is used only when
generating IL for putstatic and getstatic.

For putstatic, if it attempts to store to a static final field, it's
very likely that the store is illegal, in which case it will be treated
as unresolved. But even if the store is legal, the known object index is
useless.

For getstatic, the IL generator will use foldReliableStaticFinalField()
from TransformUtil to get the same effect anyway, so setting the known
object index here is redundant.