Handle resolved interface method lookups for default implementations
In TR_ResolvedJ9Method::getResolvedInterfaceMethod, if the the class
of the interface method is an interface class, then the resulting
resolved method was just discarded, as the interface class was not
expected to contain any interface method implementation prior to
default interface methods being introduced in Java 8. This change
removes the condition ensuring that the interface methods are not
inside an interface class, enabling this function to obtain resolved
interface methods that are default implementations.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>