Disable use of method handle for core reflection in JDK21
In JEP 416, core reflection was re-implemented with Method Handles with
the purpose of simplifying adding new language features. While analyzing
performance of workload based on large enterprise based application, I
observed that this was causing visible performance degradation.
Disabling the use of direct method handles for reflection calls while we
are working on improving the performance with direct method handles.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>