Change NeedsPeekingHeuristics to actually run prior to use
NPH was being asked if peeking is necessary right after init, prior
to any of the bytecode iteration was taking place, following which
NPH.doPeeking() could return true. This meant that NPH was not having
any impact on whether any peeking took place, whether the heuristics
determined it being useful or not. This commit reorders the
processBytecodeAndGenerateCFG phase of ECS to take place prior to
the step checking whether peeking is necessary, so that if and when
peeking is necessary, NPH can utilized to trigger ILGen with the new
function setNeedsPeekingToTrue(). Previously, _needsPeeking was set
to true if there was a param load within distance, which is now
disabled so as not to modify the ECS behaviour where NPH was not having
any impact on peeking. Only by explicitly using
NPH.setNeedsPeekingToTrue() can peeking be triggered.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>