Stash all invokedynamic and invokehandle arguments for OSR
When stashing arguments for OSR at an invokedynamic or invokehandle
bytecode instruction, the callee usually has an extra parameter (the
appendix, if it's non-null). If the call is unresolved, it expects two
extra parameters instead (the appendix and a MemberName). For the
purpose of stashing, these are all ignored because they're not on the
stack just prior to running the invoke bytecode instruction. However,
in these cases stashArgumentsForOSR() has been ignoring the top one or
two elements of the stack as well, as though the extra arguments were
already pushed. However, they haven't actually been pushed at this
point, so we would simply fail to stash the last one or two regular
arguments. Failure to stash could cause stack slots to appear falsely to
be dead, and they could therefore be left uninitialized after an OSR
transition at runtime.

This commit changes stashArgumentsForOSR() to use the stack size as-is
because the regular arguments are still at the top of the stack.

Also keep numArgsToNotStash in the J9VM_OPT_OPENJDK_METHODHANDLE #if.

(cherry picked from commit 7dd640882f5df52cf0b0042365de86dfb69fa2e8)