Set dummy resolved method flag in findOrCreateHandleMethodSymbol()
...instead of in the caller.

This method is also called from stashArgumentsForOSR(), which neglected
to set the flag. Because the flag wasn't set, the symbol reference was
available for reuse, and it would be reused even if the invokehandle
instruction became resolved before genInvokeHandle(). In that case,
genInvokeHandle() would see isUnresolved=false, so it would push only
the appendix (if non-null), but then it would get the dummy linkToStatic
symbol reference, which expects more arguments.

This flag is now set in findOrCreateHandleMethodSymbol(). There's no
reason to allow a caller to forget to set the flag.