Increment stackSlotCount for structs correctly in FFI Upcall on z/OS
This commit fixes a bug in FFI Upcall on z/OS where the counter for
stack slots in memory is not always increased when iterating through
arguments. Specifically, the counter was not being incremented when
structs were present in the argument list. It will now be incremented
correctly for every argument.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>