Prevent expanding sanity for Valhalla testing
The sanity test target is expanded to sanity.functional,sanity.openjdk.
However, for Valhalla-specific testing, such as xlinuxval, the only
test targets that are supported are sanity.functional and
extended.functional.  The sanity.openjdk is supported for the
'standard' Valhalla platform targets, such as xlinuxvalst.  This
change prevents the expansion of sanity from including sanity.openjdk
for Valhalla-specific testing.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>