Check for VarHandle's MethodType table only when necessary
A VarHandle can be initialized such that its access mode method
invocation behaves like a MethodHandle.invokeExact invocation.
We can check if this is the case by checking whether the
VarHandle.exact field was set to true. In such a case, an additional
check is necessary, where we need to verify whether
AccessDescriptor.symbolicMethodTypeExact field matches the
corresponding entry in the MethodType table, failing which the
JIT should abort trying to lookup the MethodHandleTable.

The MethodType table would only be initialized for a VarHandle
if the VarHandle.exact field was initialized as true. Therefore,
by moving the check for whether the MethodType table is valid
to occur only after we confirm that the VarHandle has
invokeExact behaviour, we prevent prematurely aborting the MH
entry lookup in getMethodHandleTableEntryIndex when the MT
table is not initialized.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>