Set log file in TR_Debug at client
When a client recompiles a method locally following a remote
compilation (enabled with the
enableJITServerFollowRemoteCompileWithLocalCompile option) it must set
up a local log file for compilation traces if necessary. Previously, it
would only set this log file explicitly in the TR::Compilation object
itself; the _file in the TR_Debug object for the compilation would not
be set. Since the TR_Debug would have been created already during the
prior remote compilation with a NULL _file, this had the effect of
suppressing the compilation traces coming from systems that used that
_file for output and not the _logFile in the TR::Compilation itself.

The file in the TR_Debug object for the compilation is now set
explicitly.

Signed-off-by: Christian Despres <despresc@ibm.com>