Always profile switches in interpreter profiler
Before this commit, the interpreter profiler was not collecting
data on switches during 'classLoadPhase'. This was done to reduce
the overhead of interpreter profiler. However, we have seen
cases where, for some methods compiled during 'classLoadPhase'
there is no profiling data for some hot switches, making the compiler
to infer the wrong block frequencies.
This commit always collects information for switches. We can afford
to do so because:
1. Typically switches are not that frequently executed compared to branches
2. When using the shared class cache (SCC), IProfiler is turned off during
   start-up, relying on profiling info stored in SCC in the cold run.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>