Fix retrieval of switch IProfiler info from SCC
The code that retrieves IProfiler info from the
shared class cache (SCC) first does a check to see
whether the data stored is valid. This check is done
against `entry->getData()`, but for switches this
always returns 0. As a consequence the JVM is never
able to retrieve switch IProfiler data from SCC.

This commit adds a new virtual call, `hasData()`,
which will be tested instead of `getData()`.
The reason we cannot use `getData()` directly, is that
the switch data uses 32 bytes, exceeding the size
returned by getData() which is only 8 bytes (for 64-bit
architectures).

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>