Updates to IProfiler branch data collection
Collection of IProfiler data is done without any form
of synchornization due to the high overhead of IProfiler.
Thus, we should try to minimize all windows of opportunities
for races to appear.
The profiling data for a branch is just a 32-bit entity which
encodes two 16-bit counters (taken and not-taken).
If incrementing one of these counters would cause an overflow,
we will divide both counters by 2 and only then increment the
desired counter. This operation is complex and cannot be done
atomically without a lock.
In order to minimize the race conditions, this commit reads the
value of the 32-bit word containing the 2 counters and performs
all the desired operations on this copy of the data before
writing it back. Since reading and writing a 32-bit aligned word
is atomic on most modern architectures, the possibility of
storing inconsistent data is decreased.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>