Adjust IProfiler locked entry assert
The TR_IProfiler::releaseAllEntries() function is used with DEBUG to
check that all entries in the IProfiler _bcHashTable are unlocked at JVM
exit. This function now compensates for a known race condition in
TR_IProfiler::findOrCreateEntry() by also reporting the number of locked
entries it encounters that were likely not a result of that race
condition, so it can be asserted instead that that total is zero.

Signed-off-by: Christian Despres <despresc@ibm.com>