Remove method TR_J9VMBase::findOrCreateClassAndDepthFlagsSymbolRef
A method named findOrCreateClassDepthAndFlagsSymbolRef was introduced to
TR_J9VMBase by a previous commit to correct the typo in the name of the
method, findOrCreateClassAndDepthFlagsSymbolRef.  That change kept the
old method due to upstream uses in J9_PROJECT_SPECIFIC code in OMR.

This change replaces a call to findOrCreateClassAndDepthFlagsSymbolRef
that was recently introduced with a call to
TR_J9VMBase::testIsClassArrayType instead, to hide the manipulation of
the classDepthAndFlags field.

As that was the last remaining use of
findOrCreateClassAndDepthFlagsSymbolRef, this change also removes the
now obsolete method from TR_J9VMBase.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>